rating_service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
mysql-connector-python==8.0.27
mysqlclient==2.1.1
requests==2.26.0
google-generativeai==0.3.0 